a2s-loader/requirements.txt
python-a2s
asyncio
clickhouse_driver